added cmds and edited couple files
added vcredist.js command, finished, added userinfo.js which is still in progress. edited purge.js, lastllog.js, purge.js, embed.json, index.js and finished commands, added error color for embed.json and added presense in the index.

+++ src/cmd/moderation/purge.js
@@ -1,6 +1,6 @@
 const { CommandType } = require('wokcommands');
 const { EmbedBuilder, ApplicationCommandOptionType } = require('discord.js');
-const { color } = require('../../cfg/embed/embed.json')
+const { color, errcolor } = require('../../cfg/embed/embed.json')
 
 module.exports = {
     description: 'Delete a certain amount of messages.',
@@ -17,16 +17,43 @@ module.exports = {
         },
     ],
 
-    callback: async ({ message, interaction, args, channel }) => {
+    callback: async ({ interaction, args, channel }) => {
         const amount = args.length ? parseInt(args.shift()) : 10
+
+        if (amount > 100) {
+            const errorembed = new EmbedBuilder()
+            .setColor(errcolor)
+            .setTitle(`You can't delete more than 100 messages at a time.`)
+
+            interaction.reply({
+                embeds: [errorembed],
+                ephemeral: true,
+            })
+
+            return
+        }
+
+        if (amount < 2) {
+            const errorembed = new EmbedBuilder()
+            .setColor(errcolor)
+            .setTitle(`You can't delete less than 2 messages at a time.`)
+
+            interaction.reply({
+                embeds: [errorembed],
+                ephemeral: true,
+            })
+
+            return
+        }
 
         const { size } = await channel.bulkDelete(amount, true)
 
         const embedmsg = new EmbedBuilder()
         .setColor(color)
+        .setTitle(`Deleted ${size} messages.`)
 
         interaction.reply({
-            content: `Deleted ${size} messages.`,
+            embeds: [embedmsg],
             ephemeral: true
         })
     }
